Switch codedeploy commands to use clients
